djackets_django/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djoser==2.1.0
psycopg2==2.9.7
django-cors-headers==4.4.0


